@@ -1,5 +1,5 @@
 import React from 'react';
-import {HashRouter as Router, Switch, Route}from 'react-router-dom'
+import {HashRouter as Router, Routes, Route}from 'react-router-dom'
 import Home from './Home'
 import Week from './Week'
 import './style/main.scss'
@@ -8,10 +8,10 @@ function App() {
   return (
     <div className="App">
       <Router>
-        <Switch>
-          <Route exact path='/' component={Home} />
-          <Route path='/week/:number' component={Week}/> 
-        </Switch>
+        <Routes>
+          <Route path='/' element={<Home/>} />
+          <Route path='/week/:number' element={<Week/>}/> 
+        </Routes>
       </Router>
     </div>
   );
